# Don't use rate limiter when listing resources in multiple chunks