@brianpursley 2020-06-24
> That's possibly correct, but that doesn't resolve my initial problem I think.

Ah, you mean wanting to be able to see the partial results before the entire request is complete?

This probably makes that less of a problem, but you're right, it does nothing to make the partial results available sooner.

I will update this PR to say it is related to the issue, but does not FIX the issue.