@apelisse 2020-06-24
> I will update this PR to say it is related to the issue, but does not FIX the issue.

I appreciate that and the work that you're doing Brian, thanks!

To explain the use-case, my initial problem happened as I was trying to remove resources from a cluster that was completely overloaded. And yet I wasn't even able to use kubectl to list the resources that I wanted to remove. I could see from the logs that SOME requests were getting results very slowly, but because of the caching, I wasn't able to see these.