@liggitt 2020-06-25
removing the ratelimiter here seems like dropping this protection in exactly the sort of place we need it