@liggitt 2020-06-25
This seems like the rate limiter working properly, I'm not sure I understand why this is considered a bug