@brianpursley 2020-06-25
> This seems like the rate limiter working properly, I'm not sure I understand why this is considered a bug

Yeah you’re probably right.  Besides you can increase chunk size to avoid throttling if it causes a problem, right?

This PR came out of a discussion in the issue where listing 100k items took an extremely long time.

I’ll go ahead and close it, since there is an already existing way to control this.