@ganga1980 2020-12-03
Hi, @brianpursley , @apelisse,  @liggitt , can you please help on understanding whether rate limiting's applies to chunked calls? if yes, what are the rate limits?  I am working on lowering the chunk-size  to best optimal performance for the client and server. Based on my understanding, rate applies only to non-chunked api calls. please correct me here if my understanding not correct.